requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.8
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary~=2.9.3
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
